Split IssueResolver class into a separate file (#8619)
