+++ tests/unit/resolver/test_resolve_issue.py
@@ -4,7 +4,7 @@ import pytest
 
 from openhands.core.config import SandboxConfig
 from openhands.events.action import CmdRunAction
-from openhands.resolver.resolve_issue import IssueResolver
+from openhands.resolver.issue_resolver import IssueResolver
 
 
 def assert_sandbox_config(
@@ -87,8 +87,8 @@ def test_setup_sandbox_config_experimental():
         assert_sandbox_config(config, runtime_container_image=None)
 
 
-@mock.patch('openhands.resolver.resolve_issue.os.getuid', return_value=0)
-@mock.patch('openhands.resolver.resolve_issue.get_unique_uid', return_value=1001)
+@mock.patch('openhands.resolver.issue_resolver.os.getuid', return_value=0)
+@mock.patch('openhands.resolver.issue_resolver.get_unique_uid', return_value=1001)
 def test_setup_sandbox_config_gitlab_ci(mock_get_unique_uid, mock_getuid):
     """Test GitLab CI specific configuration when running as root"""
     with mock.patch('openhands.__version__', 'mock'):
@@ -102,7 +102,7 @@ def test_setup_sandbox_config_gitlab_ci(mock_get_unique_uid, mock_getuid):
             assert_sandbox_config(config, local_runtime_url='http://localhost')
 
 
-@mock.patch('openhands.resolver.resolve_issue.os.getuid', return_value=1000)
+@mock.patch('openhands.resolver.issue_resolver.os.getuid', return_value=1000)
 def test_setup_sandbox_config_gitlab_ci_non_root(mock_getuid):
     """Test GitLab CI configuration when not running as root"""
     with mock.patch('openhands.__version__', 'mock'):
